article.php

+++ article.php
@@ -1,12 +1,12 @@
 <div class="container">
-    		<div class="row mb-2">
+    <div class="row mb-2">
         <div class="col-md-6">
-           <button type="button" class="btn btn-secondary mb-2" data-bs-toggle="modal" data-bs-target="#modalTambah">
-        <i class="bi bi-plus-lg"></i> Tambah Article
-    </button>
+            <button type="button" class="btn btn-secondary mb-2" data-bs-toggle="modal" data-bs-target="#modalTambah">
+                <i class="bi bi-plus-lg"></i> Tambah Article
+            </button>
         </div>
         <div class="col-md-6">
-        						<div class="input-group">
+            <div class="input-group">
                 <input type="text" id="search" class="form-control" placeholder="Cari Artikel...">
                 <span class="input-group-text">
                     <i class="bi bi-search"></i>
@@ -14,6 +14,7 @@
             </div>
         </div>
     </div>
+
     <div class="row">
         <div class="table-responsive">
             <table class="table table-hover">
@@ -27,7 +28,15 @@
                     </tr>
                 </thead>
                 <tbody id="result">
-                   
+                    <?php
+                    // 1. QUERY DATA
+                    $sql = "SELECT * FROM article ORDER BY tanggal DESC";
+                    $hasil = $conn->query($sql);
+                    $no = 1;
+
+                    // 2. LOOPING DATA
+                    while ($row = $hasil->fetch_assoc()) {
+                    ?>
                         <tr>
                             <td><?= $no++ ?></td>
                             <td>
@@ -40,98 +49,96 @@
                                 <?php
                                 if ($row["gambar"] != '') {
                                     if (file_exists('img/' . $row["gambar"] . '')) {
-                                        echo '<img src="img/' . $row["gambar"] . '" class="img-fluid" alt="Gambar Artikel">';
+                                        echo '<img src="img/' . $row["gambar"] . '" class="img-fluid" width="150" alt="Gambar Artikel">';
                                     }
                                 }
                                 ?>
                             </td>
                             <td>
-                              <a href="#" title="edit" class="badge rounded-pill text-bg-success" data-bs-toggle="modal" data-bs-target="#modalEdit<?= $row["id"] ?>"><i class="bi bi-pencil"></i></a>
-                              <a href="#" title="delete" class="badge rounded-pill text-bg-danger" data-bs-toggle="modal" data-bs-target="#modalHapus<?= $row["id"] ?>"><i class="bi bi-x-circle"></i></a>
-                            <!-- Awal Modal Edit -->
-<div class="modal fade" id="modalEdit<?= $row["id"] ?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h1 class="modal-title fs-5" id="staticBackdropLabel">Edit Article</h1>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form method="post" action="" enctype="multipart/form-data">
-                <div class="modal-body">
-                    <div class="mb-3">
-                        <label for="judul" class="form-label">Judul</label>
-                        <input type="hidden" name="id" value="<?= $row["id"] ?>">
-                        <input type="text" class="form-control" name="judul" placeholder="Tuliskan Judul Artikel" value="<?= $row["judul"] ?>" required>
-                    </div>
-                    <div class="mb-3">
-                        <label for="isi">Isi</label>
-                        <textarea class="form-control" placeholder="Tuliskan Isi Artikel" name="isi" required><?= $row["isi"] ?></textarea>
-                    </div>
-                    <div class="mb-3">
-                        <label for="gambar" class="form-label">Ganti Gambar</label>
-                        <input type="file" class="form-control" name="gambar">
-                    </div>
-                    <div class="mb-3">
-                        <label for="formGroupExampleInput3" class="form-label">Gambar Lama</label>
-                        <?php
-                        if ($row["gambar"] != '') {
-                            if (file_exists('img/' . $row["gambar"] . '')) { 
-                                echo '<br><img src="img/' . $row["gambar"] . '" class="img-fluid" alt="Gambar Artikel">';
-                            }
-                        }
-                        ?>
-                        <input type="hidden" name="gambar_lama" value="<?= $row["gambar"] ?>">
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-                    <input type="submit" value="simpan" name="simpan" class="btn btn-primary">
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
-<!-- Akhir Modal Edit -->
-
-<!-- Awal Modal Hapus -->
-<div class="modal fade" id="modalHapus<?= $row["id"] ?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
-    <div class="modal-dialog">
-        <div class="modal-content">
-            <div class="modal-header">
-                <h1 class="modal-title fs-5" id="staticBackdropLabel">Konfirmasi Hapus Article</h1>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-            </div>
-            <form method="post" action="" enctype="multipart/form-data">
-                <div class="modal-body">
-                    <div class="mb-3">
-                        <label for="formGroupExampleInput" class="form-label">Yakin akan menghapus artikel "<strong><?= $row["judul"] ?></strong>"?</label>
-                        <input type="hidden" name="id" value="<?= $row["id"] ?>">
-                        <input type="hidden" name="gambar" value="<?= $row["gambar"] ?>">
-                    </div>
-                </div>
-                <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">batal</button>
-                    <input type="submit" value="hapus" name="hapus" class="btn btn-primary">
-                </div>
-            </form>
-        </div>
-    </div>
-</div>
-<!-- Akhir Modal Hapus -->
+                                <a href="#" title="edit" class="badge rounded-pill text-bg-success" data-bs-toggle="modal" data-bs-target="#modalEdit<?= $row["id"] ?>"><i class="bi bi-pencil"></i></a>
+                                <a href="#" title="delete" class="badge rounded-pill text-bg-danger" data-bs-toggle="modal" data-bs-target="#modalHapus<?= $row["id"] ?>"><i class="bi bi-x-circle"></i></a>
                             </td>
                         </tr>
-                    <?php
+
+                        <div class="modal fade" id="modalEdit<?= $row["id"] ?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+                            <div class="modal-dialog">
+                                <div class="modal-content">
+                                    <div class="modal-header">
+                                        <h1 class="modal-title fs-5" id="staticBackdropLabel">Edit Article</h1>
+                                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                    </div>
+                                    <form method="post" action="" enctype="multipart/form-data">
+                                        <div class="modal-body">
+                                            <div class="mb-3">
+                                                <label for="judul" class="form-label">Judul</label>
+                                                <input type="hidden" name="id" value="<?= $row["id"] ?>">
+                                                <input type="text" class="form-control" name="judul" placeholder="Tuliskan Judul Artikel" value="<?= $row["judul"] ?>" required>
+                                            </div>
+                                            <div class="mb-3">
+                                                <label for="isi">Isi</label>
+                                                <textarea class="form-control" placeholder="Tuliskan Isi Artikel" name="isi" required><?= $row["isi"] ?></textarea>
+                                            </div>
+                                            <div class="mb-3">
+                                                <label for="gambar" class="form-label">Ganti Gambar</label>
+                                                <input type="file" class="form-control" name="gambar">
+                                            </div>
+                                            <div class="mb-3">
+                                                <label for="formGroupExampleInput3" class="form-label">Gambar Lama</label>
+                                                <?php
+                                                if ($row["gambar"] != '') {
+                                                    if (file_exists('img/' . $row["gambar"] . '')) {
+                                                        echo '<br><img src="img/' . $row["gambar"] . '" class="img-fluid" width="100" alt="Gambar Artikel">';
+                                                    }
+                                                }
+                                                ?>
+                                                <input type="hidden" name="gambar_lama" value="<?= $row["gambar"] ?>">
+                                            </div>
+                                        </div>
+                                        <div class="modal-footer">
+                                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                                            <input type="submit" value="simpan" name="simpan" class="btn btn-primary">
+                                        </div>
+                                    </form>
+                                </div>
+                            </div>
+                        </div>
+                        <div class="modal fade" id="modalHapus<?= $row["id"] ?>" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
+                            <div class="modal-dialog">
+                                <div class="modal-content">
+                                    <div class="modal-header">
+                                        <h1 class="modal-title fs-5" id="staticBackdropLabel">Konfirmasi Hapus Article</h1>
+                                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                                    </div>
+                                    <form method="post" action="" enctype="multipart/form-data">
+                                        <div class="modal-body">
+                                            <div class="mb-3">
+                                                <label for="formGroupExampleInput" class="form-label">Yakin akan menghapus artikel "<strong><?= $row["judul"] ?></strong>"?</label>
+                                                <input type="hidden" name="id" value="<?= $row["id"] ?>">
+                                                <input type="hidden" name="gambar" value="<?= $row["gambar"] ?>">
+                                            </div>
+                                        </div>
+                                        <div class="modal-footer">
+                                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">batal</button>
+                                            <input type="submit" value="hapus" name="hapus" class="btn btn-primary">
+                                        </div>
+                                    </form>
+                                </div>
+                            </div>
+                        </div>
+                        <?php
+                    } // AKHIR LOOPING
                     ?>
                 </tbody>
             </table>
         </div>
-     </div>
-     </div> 
- <div class="modal fade" id="modalTambah" tabindex="-1" aria-labelledby="modalTambahLabel" aria-hidden="true">
+    </div>
+</div>
+
+<div class="modal fade" id="modalTambah" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1" aria-labelledby="staticBackdropLabel" aria-hidden="true">
     <div class="modal-dialog">
         <div class="modal-content">
             <div class="modal-header">
-                <h1 class="modal-title fs-5" id="modalTambahLabel">Tambah Artikel</h1>
+                <h1 class="modal-title fs-5" id="staticBackdropLabel">Tambah Article</h1>
                 <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
             </div>
             <form method="post" action="" enctype="multipart/form-data">
@@ -145,12 +152,12 @@
                         <textarea class="form-control" placeholder="Tuliskan Isi Artikel" name="isi" required></textarea>
                     </div>
                     <div class="mb-3">
-                        <label for="gambar" class="form-label">Gambar</label>
+                        <label for="gambar" class="form-label">Pilih Gambar</label>
                         <input type="file" class="form-control" name="gambar">
                     </div>
                 </div>
                 <div class="modal-footer">
-                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Tutup</button>
+                    <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
                     <input type="submit" value="simpan" name="simpan" class="btn btn-primary">
                 </div>
             </form>
@@ -158,35 +165,28 @@
     </div>
 </div>
 <script>
-    function loadData(keyword = '') {
-        $.ajax({
-            url: "article_search.php",
-            type: "POST",
-            data: {
-                keyword: keyword
-            },
-            success: function(data) {
-                $("#result").html(data);
-            }
+    $(document).ready(function() {
+        // Event pencarian (Kode Ajax)
+        $("#search").on("keyup", function() {
+            var keyword = $(this).val();
+            $.ajax({
+                url: "article_search.php",
+                type: "POST",
+                data: {
+                    keyword: keyword
+                },
+                success: function(data) {
+                    $("#result").html(data);
+                }
+            });
         });
-    }
-
-    // load awal
-    loadData();
-
-    		// event pencarian
-    $("#search").on("keyup", function() {
-        let keyword = $(this).val();
-        if(keyword.length >=3 || keyword.lenght === 0){
-            loadData(keyword);
-        }
     });
 </script>
 
 <?php
 include "upload_foto.php";
 
-//jika tombol simpan diklik
+// LOGIKA SIMPAN (INSERT & UPDATE) DALAM SATU BLOK
 if (isset($_POST['simpan'])) {
     $judul = $_POST['judul'];
     $isi = $_POST['isi'];
@@ -195,18 +195,12 @@ if (isset($_POST['simpan'])) {
     $gambar = '';
     $nama_gambar = $_FILES['gambar']['name'];
 
-    //jika ada file baru yang dikirim  
+    // Cek jika ada file gambar baru
     if ($nama_gambar != '') {
-        //panggil function upload_foto untuk cek detail file yg diupload user
-        //function ini memiliki keluaran sebuah array yang berisi status dan message
         $cek_upload = upload_foto($_FILES["gambar"]);
-
-        //cek status upload file hasilnya true/false
         if ($cek_upload['status']) {
-            //jika true maka message berisi nama file gambar
             $gambar = $cek_upload['message'];
         } else {
-            //jika true maka message berisi pesan error, tampilkan dalam alert
             echo "<script>
                 alert('" . $cek_upload['message'] . "');
                 document.location='admin.php?page=article';
@@ -215,35 +209,26 @@ if (isset($_POST['simpan'])) {
         }
     }
 
-    		//cek apakah ada id yang dikirimkan dari form
+    // CEK APAKAH INI EDIT (Ada ID) ATAU TAMBAH (Tidak ada ID)
     if (isset($_POST['id'])) {
-        //jika ada id, lakukan update data dengan id tersebut
+        // --- LOGIKA UPDATE ---
         $id = $_POST['id'];
 
         if ($nama_gambar == '') {
-            //jika tidak ganti gambar
-            $gambar = $_POST['gambar_lama'];
+            $gambar = $_POST['gambar_lama']; // Pakai gambar lama jika tidak upload baru
         } else {
-            //jika ganti gambar, hapus gambar lama
-            unlink("img/" . $_POST['gambar_lama']);
+            // Hapus gambar lama jika upload baru
+            if (file_exists("img/" . $_POST['gambar_lama'])) {
+                unlink("img/" . $_POST['gambar_lama']);
+            }
         }
 
-        $stmt = $conn->prepare("UPDATE article 
-                                SET 
-                                judul =?,
-                                isi =?,
-                                gambar = ?,
-                                tanggal = ?,
-                                username = ?
-                                WHERE id = ?");
-
+        $stmt = $conn->prepare("UPDATE article SET judul=?, isi=?, gambar=?, tanggal=?, username=? WHERE id=?");
         $stmt->bind_param("sssssi", $judul, $isi, $gambar, $tanggal, $username, $id);
         $simpan = $stmt->execute();
     } else {
-		    //jika tidak ada id, lakukan insert data baru
-        $stmt = $conn->prepare("INSERT INTO article (judul,isi,gambar,tanggal,username)
-                                VALUES (?,?,?,?,?)");
-
+        // --- LOGIKA INSERT (TAMBAH BARU) ---
+        $stmt = $conn->prepare("INSERT INTO article (judul,isi,gambar,tanggal,username) VALUES (?,?,?,?,?)");
         $stmt->bind_param("sssss", $judul, $isi, $gambar, $tanggal, $username);
         $simpan = $stmt->execute();
     }
@@ -263,18 +248,19 @@ if (isset($_POST['simpan'])) {
     $stmt->close();
     $conn->close();
 }
- //jika tombol hapus diklik
+
+// LOGIKA HAPUS
 if (isset($_POST['hapus'])) {
     $id = $_POST['id'];
     $gambar = $_POST['gambar'];
 
     if ($gambar != '') {
-        //hapus file gambar dari folder /img
-        unlink("img/" . $gambar);
+        if (file_exists("img/" . $gambar)) {
+            unlink("img/" . $gambar);
+        }
     }
 
     $stmt = $conn->prepare("DELETE FROM article WHERE id =?");
-
     $stmt->bind_param("i", $id);
     $hapus = $stmt->execute();
